refactor: BookInfoVIew 이동 메서드 분리
