+++ BookApp/View/Search/BookSearchViewController.swift
@@ -144,14 +144,7 @@ class BookSearchViewController: UIViewController {
 extension BookSearchViewController: UICollectionViewDelegate {
     
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        let bottomSheetVC = BookInfoViewController()
-        if let sheet = bottomSheetVC.sheetPresentationController {
-            sheet.detents = [.custom(resolver: { context in
-                return context.maximumDetentValue * 0.9 })]
-            sheet.preferredCornerRadius = 20      
-        }
-        bottomSheetVC.modalPresentationStyle = .pageSheet
-        present(bottomSheetVC, animated: true, completion: nil)
+        navigateToBookInfoView()
     }
 }
